gcc: avoid casting function call results directly
Fix -Wbad-function-cast warnings by avoiding explicit casts of
function return values. Use intermediate variables or double casts
to satisfy the compiler's type checking.

In libdm-stats.c:
- log10() returns double, store in double variable before casting to size_t
- dm_stats_get_counter() returns uint64_t, store directly without cast,
  then use implicit conversion to double in arithmetic operations

In vdo_reader.c:
- le32toh() returns uint32_t, use double cast (uint32_t)(enum) instead
  of direct cast from function result to enum type

The argc parameter is part of the CMD_ARGS macro used by all command
functions, but _setgeometry doesn't actually use it (only decrements
it without checking the value). Mark it as intentionally unused.

+++ libdm/vdo/vdo_reader.c
@@ -147,7 +147,7 @@ static void _vdo_decode_header(struct vdo_header *h)
 
 static void _vdo_decode_geometry_region(struct vdo_volume_region *vr)
 {
-	vr->id = (enum vdo_volume_region_id) le32toh(vr->id);
+	vr->id = (enum vdo_volume_region_id) (uint32_t) le32toh(vr->id);
 	vr->start_block = le64toh(vr->start_block);
 }
